Refactored messaging and multiprocessing system, work in progress

+++ golismero_2/core/managers/processmanager.py
@@ -24,52 +24,61 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
-from core.main.commonstructures import Singleton
-from multiprocessing import Pool
+__all__ = ["ProcessManager", "OOPObserver"]
 
-# Serializable bootstrap function for subprocesses.
-# This is requied for Windows support, since we don't have os.fork() there.
+from core.main.commonstructures import  GlobalParams
+from core.messaging.message import Message
+from multiprocessing import Pool, Queue
+
+
+# Serializable bootstrap function to run plugins in subprocesses.
+# This is required for Windows support, since we don't have os.fork() there.
 # See: http://docs.python.org/2/library/multiprocessing.html#windows
-def bootstrap(module, clazz, func, *argv, **argd):
-    cls = __import__(module, fromlist=[clazz])
-    return getattr(cls(), func)(*argv, **argd)
+def bootstrap(context, module, clazz, init_argv, init_argd, func, argv, argd):
+    try:
+        cls = __import__(module, fromlist=[clazz])
+        instance = cls(*init_argv, **init_argd)
+        if context is not None and hasattr(instance, "_set_observer"):
+            instance._set_observer( OOPObserver(context) )
+        return getattr(instance, func)(*argv, **argd)
+    except:
+        # XXX DEBUG
+        import traceback
+        traceback.print_exc()
+
 
 #------------------------------------------------------------------------------
-class ProcessManager(Singleton):
+class ProcessManager (object):
     """
-    This class manages processes.
-
-    To run a function you must do the following::
-
-        # Definitions
-        class A():
-            def show(self, message):
-                print "hello " + message
-
-        # Main
-        if __name__ == '__main__':
-            a = A()
-            p = ProcessManager()
-            p.execute(__name__, 'A', 'show', ('world',))
+    Manages a pool of subprocesses to run plugins in them.
     """
+
 
     #----------------------------------------------------------------------
-    def __init__(self, max_processes = None, refresh_after_tasks = None):
+    def __init__(self, config):
         """Constructor.
 
-        :param max_processes: maximum number of processes to create
-        :type max_processes: int
-
-        :param refresh_after_tasks: maximum number of function calls to make before refreshing a subprocess
-        :type refresh_after_tasks: int
+        :param config: Configuration object
+        :type config: GlobalParams
         """
-        super(ProcessManager, self).__init__()
-        self.__pool = Pool(max_processes, maxtasksperchild = refresh_after_tasks)
+
+        # maximum number of processes to create
+        self.__max_processes       = getattr(config, "max_processes",       None)
+
+        # maximum number of function calls to make before refreshing a subprocess
+        self.__refresh_after_tasks = getattr(config, "refresh_after_tasks", None)
+
+        # no process pool for now...
+        self.__pool = None
+
 
     #----------------------------------------------------------------------
-    def execute(self, clazz, func, *argv, **argd):
+    def execute(self, context, module, clazz, init_argv, init_argd, func, argv, argd):
         """
-        Add a function to be executed, with its params, in a pooled process.
+        Run a plugin in a pooled process.
+
+        :param context: context for the bootstrap function
+        :type context: Context
 
         :param module: module where the class is defined
         :type module: str
@@ -77,17 +86,57 @@ class ProcessManager(Singleton):
         :param clazz: class of the plugin to run, must be serializable
         :type clazz: str
 
+        :param init_argv: positional arguments to the constructor
+        :type init_argv: tuple
+
+        :param init_argd: keyword arguments to the constructor
+        :type init_argd: dict
+
         :param func: name of the method to execute, all extra arguments to this function will be passed to it
         :type func: str
+
+        :param argv: positional arguments to the function call
+        :type argv: tuple
+
+        :param argd: keyword arguments to the function call
+        :type argd: dict
         """
-        self.__pool.apply_async(bootstrap, (clazz, func, argv, argd))
+
+        # If we have a process pool, run the plugin asynchronously
+        if self.__pool is not None:
+            return self.__pool.apply_async(bootstrap,
+                    (context,
+                     module, clazz, init_argv, init_argd,
+                     func, argv, argd))
+
+        # Otherwise just call the plugin directly
+        return bootstrap(context,
+                         module, clazz, init_argv, init_argd,
+                         func, argv, argd)
+
 
     #----------------------------------------------------------------------
     def start(self):
         """
         Start the process manager.
         """
-        pass
+
+        # If we already have a process pool, do nothing
+        if self.__pool is None:
+
+            # Are we running the plugins in multiprocessing mode?
+            if self.__max_processes is not None and self.__max_processes > 0:
+
+                # Create the process pool
+                self.__pool = Pool(self.__max_processes,
+                                   maxtasksperchild = self.__refresh_after_tasks)
+
+            # Are we running the plugins in single process mode?
+            else:
+
+                # No process pool then!
+                self.__pool = None
+
 
     #----------------------------------------------------------------------
     def stop(self, wait = False):
@@ -97,8 +146,61 @@ class ProcessManager(Singleton):
         :param wait: True to wait for the subprocesses to finish, False to kill them.
         :type wait: bool
         """
-        if wait:
-            self.__pool.close()
-        else:
-            self.__pool.terminate()
-        self.__pool.join()
+
+        # If we have a process pool...
+        if self.__pool is not None:
+
+            # Either wait patiently, or kill the damn things!
+            if wait:
+                self.__pool.close()
+            else:
+                self.__pool.terminate()
+            self.__pool.join()
+
+            # Destroy the process pool
+            self.__pool = None
+
+
+#------------------------------------------------------------------------------
+class Context (object):
+    """
+    Serializable execution context for the OOP Observer.
+    """
+
+    def __init__(self, audit_name, msg_queue):
+        self.__audit_name = audit_name
+        self.__msg_queue  = msg_queue
+
+    @property
+    def audit_name(self):
+        return self.__audit_name
+
+    @property
+    def msg_queue(self):
+        return self.__msg_queue
+
+
+#------------------------------------------------------------------------------
+class OOPObserver (object):
+    """
+    Observer that proxies messages across different processes.
+    """
+
+    def __init__(self, context):
+        """
+        :param context: Execution context for the OOP observer.
+        :type context: Context
+        """
+        super(OOPObserver, self).__init__(self)
+        self.__context = context
+
+    def send_info(self, result):
+        """
+        Send results from the plugins to the orchestrator.
+
+        :param result:
+        """
+        message = Message(message_info = result,
+                          message_type = Message.MSG_TYPE_INFO,
+                          audit_name   = self.__context.audit_name)
+        self.__context.msg_queue.put_nowait(message)
